(Issue #6) This gives me a working installed wheel.
Note that I am building the wheel and using it on the same machine.  Therefore,
the fact that the package's tests are passing does not imply that they will
pass if installed on a different machine.  For instance, a different machine
might have OpenMPI installed in a different location.  Note that not all
C++-based tests are building or passing.