(Issue #6) Get automatic builds when installing from sdist.
This is now using meson under the hood to identify the local MPI installation
and use this to build the C++ OpenBT CLTs.  These are then included in a
locally-built binary wheel and installed.

Ran the in-package tests and these are all passing.  I see four MPI processes
running at 100%.