(Issue #6) Install latest meson into test venv from PyPI.
The meson documentation suggests that this should be preferred over installation
via package manager since the rapid development of meson will likley result in
only significantly old versions available through the managers.  This is the
case for Ubuntu 20.04 and 22.04 at the very least.  Also, it looks like meson
might now correctly find MPICH installations as of the latest versions, v1.6.0,
which is too new even for Ubuntu 24.04.